새로운 플레이어가 룸에 들어오거나 나가더라도 룸 리스트 UI의 플레이어 수 정보가 자동으로 갱신되지 않는 문제 해결
드롭다운 문제였음

+++ Assets/02_Scripts/Network/LobbyManager.cs
@@ -141,7 +141,7 @@ public class LobbyManager : MonoBehaviourPunCallbacks
     {
         // 방 옵션 설정
         RoomOptions options = new RoomOptions();
-        options.MaxPlayers = (byte)(maxPlayersDropdown.value + 2); // 최소 2명부터
+        options.MaxPlayers = maxPlayersDropdown.value + 2; // 최소 2명부터
         options.CustomRoomProperties = new ExitGames.Client.Photon.Hashtable()
         {
             {"password", roomPasswordInput.text}, // 방 비밀번호 설정
